Add bin/bluserawwatch.jl script
This executable script is intended to be run as a long running process.
Typically it would be managed by process manager such as systemd,
circusd, or monit.  The normal way to run this script is actually via
bash rather than julia.  This allows determination of the project
directory from the location of the script itself.  For this to work
however, the script itself must be within the directory structure of
this package and then symlinked to from a more "normal" spot (e.g.
/usr/local/bin).  Because using Pkg.add() to add the package will put it
in a tricky to find directory, it is currently recommended to Pkg.dev()
this package.  This will put the script in the bin subdirectory of
$HOME/.julia/dev/BlueRawWatch, making it easier to symlink to.