Add example circusd config file
This example circusd config file defines a watcher named
"bluse_raw_watcher" and is designed to run /home/obs/bin/bluserawwatc.jl
which should be a symlink to the bin/bluserawwatch.jl file in this repo.
The example directory that gets watched is "/buf0ro", but for general
use this should probably come from a circus environment variable.